+++ src/browseraction/__tests__/logic/load.spec.ts
@@ -1,9 +1,17 @@
 import browser from 'webextension-polyfill'
 import { afterEach, describe, expect, it, vi } from 'vitest'
-import { getTabCount, getURLsFromText, loadSites } from '@/browseraction/components/logic/load'
+import { getTabCount, getURLsFromText, loadSites, NEW_TAB_GROUP_ID, NO_TAB_GROUP_ID } from '@/browseraction/components/logic/load'
+
+const MOCK_TAB_GROUP_ID = 42
 
 vi.mock('webextension-polyfill', () => ({
-  default: { tabs: { create: vi.fn() }, runtime: { getURL: (val: string) => val } }
+  default: {
+    tabs: { 
+      create: vi.fn(), 
+      group: vi.fn() 
+    },
+    runtime: { getURL: (val: string) => val }
+  }
 }))
 
 const url1 = 'https://test.de'
@@ -16,7 +24,7 @@ describe('load tabs', () => {
   })
 
   it('loads tabs in sequence', async () => {
-    loadSites(urlList, false, false, false, false)
+    await loadSites(urlList, false, false, false, false, null)
 
     expect(browser.tabs.create).toHaveBeenNthCalledWith(1, {
       url: url1,
@@ -27,10 +35,11 @@ describe('load tabs', () => {
       active: false
     })
     expect(browser.tabs.create).toHaveBeenCalledTimes(2)
+    expect(browser.tabs.group).not.toHaveBeenCalled()
   })
 
   it('lazy loads tabs', async () => {
-    loadSites(urlList, true, false, false, false)
+    await loadSites(urlList, true, false, false, false, null)
 
     expect(browser.tabs.create).toHaveBeenCalledWith({
       url: 'lazyloading.html#' + url1,
@@ -44,13 +53,13 @@ describe('load tabs', () => {
   })
 
   it('loads tabs in random order', async () => {
-    loadSites(urlList, false, true, false, false)
+    await loadSites(urlList, false, true, false, false, null)
 
     expect(browser.tabs.create).toHaveBeenCalledTimes(2)
   })
 
   it('loads tabs in reverse order', async () => {
-    loadSites(urlList, false, false, true, false)
+    await loadSites(urlList, false, false, true, false, null)
 
     expect(browser.tabs.create).toHaveBeenNthCalledWith(1, {
       url: url2,
@@ -64,7 +73,7 @@ describe('load tabs', () => {
   })
 
   it('loads tabs and deduplicate', async () => {
-    loadSites(`${urlList}\n${urlList}\n${urlList}\n${urlList}`, false, false, false, true)
+    await loadSites(`${urlList}\n${urlList}\n${urlList}\n${urlList}`, false, false, false, true, null)
 
     expect(browser.tabs.create).toHaveBeenNthCalledWith(1, {
       url: url1,
@@ -78,7 +87,7 @@ describe('load tabs', () => {
   })
 
   it('appends http protocol if protocol does not exist', async () => {
-    loadSites('test.de', false, false, true, false)
+    await loadSites('test.de', false, false, true, false, null)
 
     expect(browser.tabs.create).toHaveBeenNthCalledWith(1, {
       url: 'http://test.de',
@@ -106,5 +115,26 @@ describe('load tabs', () => {
       url2
     ])
     expect(getURLsFromText(`\n\n\n${urlList}\n\n\n${urlList}\n\n`, true)).toEqual([url1, url2])
+  })
+
+  it('loads tabs without tab group', async () => {
+    await loadSites(urlList, false, false, false, false, NO_TAB_GROUP_ID)
+
+    expect(browser.tabs.create).toHaveBeenCalledTimes(2)
+    expect(browser.tabs.group).not.toHaveBeenCalled()
+  })
+
+  it('loads tabs to new tab group', async () => {
+    await loadSites(urlList, false, false, false, false, NEW_TAB_GROUP_ID)
+
+    expect(browser.tabs.create).toHaveBeenCalledTimes(2)
+    expect(browser.tabs.group).toBeCalledWith({ tabIds: [-1, -1] })
+  })
+
+  it('loads tabs to existing tab group', async () => {
+    await loadSites(urlList, false, false, false, false, MOCK_TAB_GROUP_ID)
+
+    expect(browser.tabs.create).toHaveBeenCalledTimes(2)
+    expect(browser.tabs.group).toBeCalledWith({ tabIds: [-1, -1], groupId: MOCK_TAB_GROUP_ID })
   })
 })